ENH : Rework the ball removing the game deps and apply actor model using channel
